moved legacy functions to convert_stats.py
